pyproject.toml
[project]
name = "a-copy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.2.6",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.1",
    "sentence-transformers>=5.1.2",
]
